[WIP] Serialize anonymous objects
TODO: discuss why skipping properties and fields that are string or value types and are read-only are skipped. No tests fail but there's certainly a reason why those are skipped in the first place. Probably it is for the dynamic setter and it should not be skipped for the dynamic getter?